Update Pentest Scripts Instructions
TIn this commit, I have:

T- Improved the clarity of the instructions within all pentest scripts. The scripts now provide a more precise description on how they are supposed to be run, thereby improving user experience and making them easier to use.

TThis update does not change any of the functional elements of the scripts, but it does significantly improve their documentation and usability.